ARC: ioremap_nocache optimization
On ARC700, top 256MB of address space (0xC000_0000 to 0xFFFF_FFFF) is
inherently uncached. Thus any ioremap_nocache() requests in that range
can simply return back the addr w/o need for any MMU mappings, reducing
pressure on TLB.

Also, the vanilla ioremap() now defaults to uncached semantics - this is
to workaround the braindead drivers which don't call the correct API.

The hooks for doing cached access are left in vanilla ioremap(), just in
case, but for now are effectively dead.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>