ARC: SMP resurrect #29: [plat-arcfpga] userspace prints locking bootup
This happens occasionally, when init task starts to run on CPU1, making
it's STDOUT open call land in ARC serial driver. The request_irq() call
is thus made from CPU1, leading to low level IRQ unmasking on CPU1 only,
whereas the UART IRQ is hardwired to CPU0. Thus despite the driver being
SMP re-entrant, the driver Tx locks up because TX_DONE interrupts are
asserted for CPU0, which is just not listening for them.

The workaround for now, is to unconditionally enable UART IRQ on it's
"home-cpu". Need to find the "right" way.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>