ARC: Big Endian Support
1. Obvious mods to Kconfig, Makefile
2. String routines adjustments for byte stragglers (copied from newlib)
3. The Build Config Regs (encoded as bitfields) need to be handled (see
   note below)
4. ELF loader needs minor adj
5. byteorder.h to have right version of generic header
6. jiffies_64 needs to be be flipped around

Note on Aux Reg layout changes for BE/LE
========================================
The Aux Regs layout is same bit-by-bit in both BE/LE modes.
However when casted as a bitfield encoded "C" struct, gcc treats it as
memory, generating different code for BE/LE. Thus we need different
definitions for the 2 cases.

However when manually "carving" the value for a Aux, no special handling
of BE is needed because of the very same propery.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>