ARC: signal cleanups #1: kludge in batch copying of pt_regs fixed
when preparing to call user space signal handler, kernel needs to
safe-keep the orig user mode regs in (sigframe->ucontext->sigcontext).
Instead of doing an itemised copy, as an ARC optimization, we do "bulk"
__copy_to_user(). However the layout is not exactly same as pt_regs
so we had to resort to doing a kludgy way of copy which doesn't look
very pretty.

So now we make it exactly same as pt_regs.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>