ARC: ===== OS ABI v2 ===== asm-generic/{fcntl,stat,posix_types}
Since the ABI is changeing due to unistd, we might as well move to
generic versions of these headers althoug this requires corresponding
work in uClibc as well.

stat in particular is interesting as it will allow us to make the stat
family of syscalls a simple pass-thru w/o need for any itermised copy
in userspace, hopefully speeding it up.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>