ARC: OS ABI v2 verification (elf hdr arch specific e_flags based)
To make the ABI transition less painful, added ABI verification scheme.
Essentially the pairing GNU ld now by-default adds a ABI ver tag in ELF
hdr (in the machine specific e_flags placeholder) which is checked by
kernel.
An absence implies toolchain not built with newer ld (and hence implies
that the rest of tool-chain - uClibc/gdbserver in  particular are not
adhering to the new asm-generic based syscall interface.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>