ARC: [plat-haps51] dw_pgu2fb: Use generic functions for fillrect etc.
fb_fillrect, fb_copyarea and fb_imageblit are required functions for a
frame buffer driver. They cannot be NULL, as this will cause crashes when
enabling certain kernel options (such as fbcon). This patch makes sure
that these function pointers are not NULL, but point to the generic
(unaccelerated) implementations.

This fixes STAR 9000575139.

Signed-off-by: Mischa Jonker <mischa.jonker@synopsys.com>