ARC: Makesure devtmpfs is present if building for initramfs [UPDATED]
Our initramfs now relies on devtmpfs and w/o it init just panics w/o a
clue of what's going on.

[vgupta: elaborate comment added

Noam figured out the Kconfig magic which I was tripping on !
Now Kbuild ensures that DEVTMPFS/HOTPLUG is auto-selected for INITRAMFS
We don't need the #error hack which requires manual selection.

Plausibly we could have also set DEVTMPFS_MOUNT and/or UEVENT_HELPER but
that's not apt for Kconfig, it belongs to platform speific defconfig.

kernel "code" (Kconfig) is not supposed to define policy (e.g. whether we
want to use mdev/udev etc - even if we only have mdev). That is best
done by user which means in a defconfig and/or in userland startup
scripts. The origi motivation for this change was to solve the common
annoyance where kernel with initramfs (which in turn relies on devtmps)
didn't have DEVTMPS, causing "init" to panic. We only solve that -
although that's also on the borderline of policy argument.]

Signed-off-by: Noam Camus <noamc@ezchip.com>
Signed-off-by: Vineet Gupta <vgupta@synopsys.com>