ARC: [plat-arcfpga] Enable UART to be usable when module
System boots (w/o any output on console) and then via telnet session we
can insmod serial_core.ko and arc_uart.ko
Now a write to UART should work, but it didn't.

echo "Hello" > /dev/ttyARC0

Actually because of CONFIG_SERIAL_ARC=m, platform code was not adding
uart device (because of ifdef CONFIG_SERIAL_ARC) which needs to be
CONFIG_SERIAL_ARC_MODULE

Note that as of now, UART as module ONLY works with static device nodes
and NOT with dynamic devtmpfs based nodes. This still needs to be
investigated.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>