ARC: Page Table size calculation bug
We would cap the table size to 1 page, which broke with experimental
test of 8:11:13 table walk coupled with SASID which doubles the space
needed.

In general although it is good to stive for 1 page per table, with
variable page size and such, we could end up with bigger table.

To really enforce the one page limit, the addr split must also do
calculations and do a compile time #warn

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>