ARC: [plat-aml8626x] Move AML NAND driver to drivers/mtd/nand
This is required to achieve proper link order. If you would like to
have UBIFS as rootfs, the NAND driver needs to be initialized before
UBI. However, NAND drivers need to be initialized after the generic MTD
driver framework. So, the initcalls for the NAND driver need to be
between MTD core and UBI layer. As the order of initcalls is determined
by the link order, the driver had to be moved from staging to
drivers/mtd/nand.