ARC: [plat-aml8626x] Don't initialize OTG_A device when !CONFIG_USB_GADGET
The dwc_otg driver is slow at initializing hardware (about half a second
for every controller). The AML8626X chip has two controllers, one is
connected to a USB host connector (A), the other is connected to a Mini-USB
OTG connector. If USB_GADGET is disabled, we assume this second connector
is not used, and we don't initialize this instance of dwc_otg block. This
saves half a second boot time.