ARC: [Review/signals] Tracer poke to Callee-regs were lost
Before going off into do_signal( ) we save the user-mode callee regs
(as they are not saved by default as part of pt_regs). This is to make
sure that that a Tracer (if tracing related signal) is able to do likes
of PEEKUSR(callee-reg).

However in return path we were simply discarding the user-mode callee
regs, which would break e POKEUSR(callee-reg) from a tracer.

Thanks to Al Viro for pointing this potential issue.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>