ARC: [Review] Multi-platform image #1: Kconfig enablement
Older build system only supported one platform(soc) to build at a time
and further only board could be built for that platform.
This required choice .... endchoice blocks for each of these.

So the first step towards multi-platform (and multi-board) builds it to
allow build system to do that.
So choice .. endchoice =>> menu .. endmenu

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>