ARC: ELF_NGREG was short of one register
Newer gdb/gdbserver use the ptrace regset interface for getting
user-space reg-file which in turn uses ELF_NGREG for determinign the max
request size.

Ideally it needs to mirror user_regs_struct struct, but it was not. This
would cause a ptrace request for offsetof(stop_pc) to fail.

This worked so far since ELF_NGREG was only used for core dumps where it
stop_pc was in fact not required (used primarily by gdb).

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>